Made Change Currency on front-end

+++ web/src/main/java/com/siemens/web/dto/CurrencyExchangeDTO.java
@@ -1,9 +1,6 @@
 package com.siemens.web.dto;
 
 import lombok.*;
-import org.joda.time.DateTime;
-
-import java.io.Serializable;
 
 @Setter
 @AllArgsConstructor
@@ -14,7 +11,7 @@ public class CurrencyExchangeDTO extends BaseDTO {
 
     private int currencyid1;
     private int currencyid2;
-    private int factor;
+    private Double factor;
     private String dateOfExchange;
 
     @Override
